backend/pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI backend for the full-stack application"
authors = [
    {name = "Shrek Wang", email = "shrek.wang@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.113.1",
    "uvicorn>=0.30.5",
    "pydantic[email]>=2.9.1",
    "pydantic-settings>=2.4.1",
    "sqlalchemy>=2.0.33",
    "alembic>=1.13.2",
    "psycopg[binary]>=3.2.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "celery>=5.4.0",
    "redis>=5.1.1",
    "flower>=2.0.1",
    "langchain>=0.3.4",
    "asyncpg>=0.30.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.2.0",
    "slowapi>=0.1.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=0.23.8",
    "httpx>=0.27.0",
    "greenlet>=3.2.3",
    "pytest-env>=1.1.5",
    "ruff>=0.12.12"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]
asyncio_mode = "auto"

[tool.uv]
# uv 特定配置
dev-dependencies = [
    "pytest>=8.4.0",
    "pytest-asyncio>=0.23.8", 
    "httpx>=0.27.0",
    "greenlet>=3.2.3",
    "pytest-env>=1.1.5",
    "ruff>=0.12.12"
]